Update sim to handle many-body Models
By removing the creation of the list of cartesian products of all
shapes and instead relying on a 2D kernel over the shapes themselves,
we introduce a small runtime performance loss in exchange for the
ability to handle thousands of bodies without quadratic memory requirements.

This commit is *highly* experimental and presented as-is at the point of
first functionality. Over the coming week it will be improved and made
more stable.

Signed-off-by: June Knauth (Opensuit Labs) <j@opensuit.space>